refactor: rename fake iq source module
