+++ rigflow-server/src/source/mod.rs
@@ -1,6 +1,6 @@
 use num_complex::Complex32;
 
-pub mod fake_iq;
+pub mod fake;
 pub mod factory;
 pub mod rtlsdr;
 pub mod wav;
